Consolidate auth modules to remove duplication
